Switch server to `EventEmitter`
Use `EventEmitter` instead of handlers and initializers to simplify logic.
