+++ src/bidiTab/Transport.ts
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License. *
  */
-import type {BidiTransport} from '../bidiMapper/bidiMapper.js';
+import type {BidiTransport} from '../bidiMapper/BidiMapper';
 import {
   type ChromiumBidi,
   ErrorCode,
